fixed modulazation errors.
