+++ utils/form.js
@@ -1,0 +1,44 @@
+import renderToDOM from './render';
+
+import studentsOnDom from './studentsOnDom';
+import createId from './creatId';
+import { houses, students } from './sample_data/data';
+
+const sortStudent = (e) => {
+  e.preventDefault();
+  const sortingHat = houses[Math.floor(Math.random() * houses.length)];
+
+  if (e.target.id === 'sorting') {
+    const student = document.querySelector('#student-name');
+
+    students.push({
+      id: createId(students),
+      name: student.value,
+      house: sortingHat.house,
+      crest: sortingHat.crest
+    });
+
+    student.value = '';
+    studentsOnDom('#students', students);
+  }
+};
+
+const form = () => {
+  const domString = `<form id="sorting" class="d-flex flex-column form-floating ">
+    <input
+    type="text"
+    class="form-control mb-1"
+    id="student-name"
+    placeholder="Enter a name"
+    required
+  />
+  <label for="floatingInputValue">Name to be sorted</label>
+<button type="submit" class="btn btn-success">Get Sorted!</button>
+</form>`;
+
+  renderToDOM('#form-container', domString);
+
+  document.querySelector('#sorting').addEventListener('submit', sortStudent);
+};
+
+export { form, sortStudent };
